버그 수정
실서버에 올리니 버그가 몇 개 생겨서 수정하였다.
댓글 수정시 비밀글 체크 안되는 현상 수정
댓글에 널포인트익셉션 버그 해결
게시글 작성시 널포인트익셉션 해결

+++ src/main/java/com/hansol/board/post/service/PostServiceImpl.java
@@ -124,15 +124,17 @@ public class PostServiceImpl implements PostService {
     public void remove(Long id, String password) {
         PostEntity findPost = postRepository.findById(id);
         List<Attachment> files = findPost.getAttachments();
-        for (Attachment file : files) {
-            String filePath = FILE_PATH + findPost.getCode() + "/" + file.getSavedFileName();
-            File saved = new File(filePath);
-            saved.delete();
-            String thumbnailPath = null;
-            if (StringUtils.hasText(file.getThumbnail())) {
-                thumbnailPath = FILE_PATH + findPost.getCode() + "/" + file.getThumbnail();
-                File thumbnail = new File(thumbnailPath);
-                thumbnail.delete();
+        if(files != null){
+            for (Attachment file : files) {
+                String filePath = FILE_PATH + findPost.getCode() + "/" + file.getSavedFileName();
+                File saved = new File(filePath);
+                saved.delete();
+                String thumbnailPath = null;
+                if (StringUtils.hasText(file.getThumbnail())) {
+                    thumbnailPath = FILE_PATH + findPost.getCode() + "/" + file.getThumbnail();
+                    File thumbnail = new File(thumbnailPath);
+                    thumbnail.delete();
+                }
             }
         }
         postRepository.remove(id, password);
